+++ packages/contracts/contracts/Interfaces/ICollSurplusPool.sol
@@ -12,8 +12,8 @@ interface ICollSurplusPool {
     event ActivePoolAddressChanged(address _newActivePoolAddress);
     event CollateralAddressChanged(address _newCollateralAddress);
 
-    event CollBalanceUpdated(address indexed _account, uint _newBalance);
-    event CollateralSent(address _to, uint _amount);
+    event CollBalanceUpdated(address indexed _account, uint256 _newBalance);
+    event CollateralSent(address _to, uint256 _amount);
 
     // --- Contract setters ---
 
@@ -28,7 +28,7 @@ interface ICollSurplusPool {
 
     function getCollateral(address _account) external view returns (uint);
 
-    function accountSurplus(address _account, uint _amount) external;
+    function accountSurplus(address _account, uint256 _amount) external;
 
     function claimColl(address _account) external;
