+++ packages/contracts/contracts/Interfaces/IDefaultPool.sol
@@ -8,11 +8,11 @@ import "./IPool.sol";
 interface IDefaultPool is IPool {
     // --- Events ---
     event TroveManagerAddressChanged(address _newTroveManagerAddress);
-    event DefaultPoolTHUSDDebtUpdated(uint _THUSDDebt);
-    event DefaultPoolCollateralBalanceUpdated(uint _collateral);
+    event DefaultPoolTHUSDDebtUpdated(uint256 _THUSDDebt);
+    event DefaultPoolCollateralBalanceUpdated(uint256 _collateral);
     event CollateralAddressChanged(address _newCollateralAddress);
 
     // --- Functions ---
-    function sendCollateralToActivePool(uint _amount) external;
+    function sendCollateralToActivePool(uint256 _amount) external;
     function updateCollateralBalance(uint256 _amount) external;
 }
